Moved the Qute related endpoints to its own class Hide Qute related endpoints from the OpenAPI Schema Added some OpenAPI annotations Expose swagger-ui in PROD Signed-off-by:Phillip Kruger <phillip.kruger@gmail.com>
