+++ .gitignore
@@ -37,3 +37,4 @@ nb-configuration.xml
 
 # Local environment
 .env
+/nbactions.xml
