+++ src/main/resources/application.properties
@@ -33,3 +33,15 @@ quarkus.kubernetes-client.trust-certs=true
 
 quarkus.vertx.max-event-loop-execute-time=60S
 quarkus.vertx.warning-exception-time=60S
+
+# OpenAPI
+quarkus.swagger-ui.always-include=true
+mp.openapi.extensions.smallrye.info.title=Quarkus Extension Registry API
+%dev.mp.openapi.extensions.smallrye.info.title=Quarkus Extension Registry API (dev)
+%test.mp.openapi.extensions.smallrye.info.title=Quarkus Extension Registry API (test)
+mp.openapi.extensions.smallrye.info.version=1.0.0
+mp.openapi.extensions.smallrye.info.description=Catalog of Quarkus extensions
+mp.openapi.extensions.smallrye.info.contact.email=quarkus-dev@googlegroups.com
+mp.openapi.extensions.smallrye.info.contact.name=Quarkus team
+mp.openapi.extensions.smallrye.info.contact.url=https://github.com/quarkusio/registry.quarkus.io
+quarkus.swagger-ui.title=Quarkus Extension Registry API
